# Web App Project/sportspro/server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SportsPro API",
    "version": "1.0.0",
    "description": "A simple API to perform CRUD functions for sports management software"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/",
      "description": "Version 1"
    }
  ],
  "paths": {
...
    "/customers/login": {
      "post": {
        "summary": "Log in as a customer",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "userpass123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully logged in"
          },
          "401": {
            "description": "Invalid credentials or unauthorized"
          }
        }
      }
    },
    "/customers": {
      "get": {
        "summary": "Get all customers",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        }
      },
      "post": {
        "summary": "Create a new customer",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstname",
                  "lastname",
                  "address",
                  "city",
                  "state",
                  "postalcode",
                  "countrycode",
                  "phone",
                  "email",
                  "password"
                ],
                "properties": {
                  "firstname": {
                    "type": "string",
                    "example": "Jane"
                  },
                  "lastname": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "address": {
                    "type": "string",
                    "example": "123 Main Street"
                  },
                  "city": {
                    "type": "string",
                    "example": "Brisbane"
                  },
                  "state": {
                    "type": "string",
                    "example": "QLD"
                  },
                  "postalcode": {
                    "type": "string",
                    "example": 4000
                  },
                  "countrycode": {
                    "type": "string",
                    "example": "AU"
                  },
                  "phone": {
                    "type": "string",
                    "example": "0412345678"
                  },
                  "email": {
                    "type": "string",
                    "example": "jane.doe@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "securepassword"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/customers/{customerid}": {
      "get": {
        "summary": "Get a customer by ID",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "customerid",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the customer"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer found"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      },
      "put": {
        "summary": "Update a customer's information",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "customerid",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the customer"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstname": {
                    "type": "string"
                  },
                  "lastname": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "postalcode": {
                    "type": "string"
                  },
                  "countrycode": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated"
          },
          "400": {
            "description": "Validation or input error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
...
  },
...
}